requirements.txt
requests==2.22.0
APScheduler==3.6.3
SQLAlchemy==1.3.12
beautifulsoup4==4.9.1
python-dotenv==0.14.0
psycopg2-binary==2.8.5